Issue 3710 (control-panel: error when creating an new base dn with automatically generated data)
Although it is marked as a control panel issue, the bug is on the core server and applies also to the import-ldif command-line.

The code is trying to build the path to the MakeLDIF resource directory by using the server root (install root) instead of the instance root. The MakeLDIF resource directory is located under <root>/config/MakeLDIF and so the instance root must be used.



git-svn-id: https://svn.forgerock.org/opendj/branches/b1.2@4834 41b1ffd8-f28e-4786-ab96-9950f0a78031