Fix for issue 3683
When replication replays a DELETE on an entry that has children,
the replication conflict resolution code incorrectly assumes that
this is a conflict.
This is not correct because it is also possible that the delete
was actually a subtree delete.

This can be fixed by checking the creation date of the child entry.
If it was created before the delete then this is not a conflict and the
delete must be replayed as a subtree delete.  : Subtree delete only deletes parent on replica

git-svn-id: https://svn.forgerock.org/opendj/branches/b1.2@4865 41b1ffd8-f28e-4786-ab96-9950f0a78031