Fix for issue 4020 on 2.0 (Help links in control panel cannot be customized)
The fix consists on using org.opends.server.util.ReleaseDefinition to have some customization in the product.  In this proposal the code checks for the existence of the class org.opends.server.util.ReleaseDefinition and tries to get some information out of it.  If nothing is found, the current behavior is kept.

The products wanting to have a customized help link will have to add a org.opends.quicksetup.util.ProductCustomization class in their delivery (for instance in the resources.jar jar file under the resources directory) with the following members:

public static String ADMINISTRATION_GUIDE_URL = <the admin guide URL>
public static String WIKI_MAIN_PAGE_URL = <the wiki main page URL>

In some cases (like the welcome panel of the setup and upgrade) we include on the labels HTTP URLs to the OpenDS online help (which is what issue 4020 is about).  With the following proposal, some of these labels can be modified.  The list of labels that can be changed is (the field names in org.opends.server.util.ReleaseDefinition correspond to the IDs of the Messages that can be modified):

INFO_CONTROL_PANEL_TITLE -> The title of the control panel dialog.
INFO_FRAME_UNINSTALL_TITLE -> The title of the uninstall dialog.
INFO_FRAME_UPGRADE_TITLE -> The title of the upgrade dialog.
INFO_FRAME_INSTALL_TITLE -> The title of the install dialog.
INFO_UPGRADE_WELCOME_PANEL_WEBSTART_INSTRUCTIONS -> The welcome text 
displayed in the webstart version of the upgrader.
INFO_INSTALLANDUPGRADE_WELCOME_PANEL_INSTRUCTIONS -> The welcome text 
displayed in the webstart version of the installer and upgrader (which 
is currently used when the user chooses the Java Web Start installer in 
the opends.org web page).
INFO_WELCOME_PANEL_WEBSTART_INSTRUCTIONS -> The welcome text displayed 
in the webstart version of the setup.
INFO_WELCOME_PANEL_OFFLINE_INSTRUCTIONS -> The welcome text displayed in 
the offline version of the upgrader (the one that is displayed when the 
user launches the setup command-line).


The code to check whether the ReleaseDefinition class exists or not has been refactored in the org.opends.quicksetup.util.Utils class.



git-svn-id: https://svn.forgerock.org/opendj/branches/b2.0@5404 41b1ffd8-f28e-4786-ab96-9950f0a78031